秘密鍵の画面表示隠蔽化

+++ lib/screen/config.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:shriken/class/encrypt.dart'; // EncryptManagerのインポート
+import 'package:shriken/class/encrypt.dart';
 
 class AppConfig extends StatefulWidget {
   @override
@@ -8,9 +8,10 @@ class AppConfig extends StatefulWidget {
 
 class _PageState extends State<AppConfig> {
   final _encryptManager = EncryptManager();
+  final _formKey = GlobalKey<FormState>();
+  bool _isObscured = true;
+
   final TextEditingController _nsec = TextEditingController();
-  // final TextEditingController _settingController2 = TextEditingController();
-  // final TextEditingController _settingController3 = TextEditingController();
 
   @override
   void initState() {
@@ -31,24 +32,24 @@ class _PageState extends State<AppConfig> {
   @override
   void dispose() {
     _nsec.dispose();
-    // _settingController2.dispose();
-    // _settingController3.dispose();
     super.dispose();
   }
 
   void _updateSettings() async {
-    String nsec = _nsec.text;
-    if ( nsec.isNotEmpty) {
-      print("seve nsec");
-      await _encryptManager.saveItem('nsec', nsec);
-    } else {
-      print("remove nsec");
-      await _encryptManager.deleteItem('nsec');
-    }
+    if (_formKey.currentState!.validate()) {
+      String nsec = _nsec.text;
+      if (nsec.isNotEmpty) {
+        print("seve nsec");
+        await _encryptManager.saveItem('nsec', nsec);
+      } else {
+        print("remove nsec");
+        await _encryptManager.deleteItem('nsec');
+      }
 
-    ScaffoldMessenger.of(context).showSnackBar(
-      SnackBar(content: Text('done.')),
-    );
+      ScaffoldMessenger.of(context).showSnackBar(
+        SnackBar(content: Text('done.')),
+      );
+    }
   }
 
   @override
@@ -58,39 +59,45 @@ class _PageState extends State<AppConfig> {
         title: Text('Settings'),
       ),
       body: SingleChildScrollView(
+          child: Form(
+        key: _formKey,
         child: Column(
           children: <Widget>[
             Padding(
               padding: const EdgeInsets.all(8.0),
-              child: TextField(
+              child: TextFormField(
                 controller: _nsec,
-                decoration: InputDecoration(labelText: 'nsec'),
-                maxLines: null, // 長いテキスト用
+                decoration: InputDecoration(
+                  labelText: 'nsec',
+                  suffixIcon: IconButton(
+                    icon: Icon(
+                      _isObscured ? Icons.visibility : Icons.visibility_off,
+                    ),
+                    onPressed: () {
+                      setState(() {
+                        _isObscured = !_isObscured;
+                      });
+                    },
+                  ),
+                ),
+                obscureText: _isObscured,
+                validator: (value) {
+                  if (value == null ||
+                      !value.startsWith('nsec1') ||
+                      value.length != 63) {
+                    return 'Value must start with "nsec1" and be 63 characters long';
+                  }
+                  return null;
+                },
               ),
             ),
-            // Padding(
-            //   padding: const EdgeInsets.all(8.0),
-            //   child: TextField(
-            //     controller: _settingController2,
-            //     decoration: InputDecoration(labelText: 'Setting Title 2'),
-            //     maxLines: null, // 長いテキスト用
-            //   ),
-            // ),
-            // Padding(
-            //   padding: const EdgeInsets.all(8.0),
-            //   child: TextField(
-            //     controller: _settingController3,
-            //     decoration: InputDecoration(labelText: 'Setting Title 3'),
-            //     maxLines: null, // 長いテキスト用
-            //   ),
-            // ),
             ElevatedButton(
               onPressed: _updateSettings,
               child: Text('Save Config'),
             ),
           ],
         ),
-      ),
+      )),
     );
   }
 }
